test(github): move live-network tests to integration build tag
TestFetchKeys_RealUser and TestFetchKeys_InvalidUser make live HTTP
calls to github.com. In restricted-network environments these fail
with TLS/connectivity errors rather than skipping cleanly.

Move both functions to github_integration_test.go with
  //go:build integration
so they are excluded from the default `go test ./...` run and only
run when explicitly opted in:
  go test -tags integration ./internal/github/...

The KNUCKLE_INTEGRATION=1 env-var guard is removed; the build tag is
the canonical Go mechanism for opt-in integration tests.

No coverage loss — TestClient_FetchKeys_WithTestServer (httptest) covers
the same code paths.

Closes #745
