+++ internal/github/github_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"strings"
 	"testing"
 )
@@ -14,37 +13,6 @@ func TestFetchKeys_EmptyUsername(t *testing.T) {
 	_, err := FetchKeys("")
 	if err == nil {
 		t.Fatal("expected error for empty username")
-	}
-}
-
-func TestFetchKeys_InvalidUser(t *testing.T) {
-	if os.Getenv("KNUCKLE_INTEGRATION") != "1" {
-		t.Skip("skipping live network test; set KNUCKLE_INTEGRATION=1 to run")
-	}
-	_, err := FetchKeys("this-user-definitely-does-not-exist-xyzzy-99999")
-	if err == nil {
-		t.Fatal("expected error for nonexistent user")
-	}
-}
-
-func TestFetchKeys_RealUser(t *testing.T) {
-	if os.Getenv("KNUCKLE_INTEGRATION") != "1" {
-		t.Skip("skipping live network test; set KNUCKLE_INTEGRATION=1 to run")
-	}
-	if testing.Short() {
-		t.Skip("skipping network test in short mode")
-	}
-	keys, err := FetchKeys("castrojo")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if len(keys) == 0 {
-		t.Fatal("expected at least one key")
-	}
-	for _, k := range keys {
-		if !hasValidPrefix(k) {
-			t.Errorf("key doesn't look like SSH key: %s", k[:40])
-		}
 	}
 }
